Dep-eject app.views.
