+++ lib/depject/index.js
@@ -21,8 +21,7 @@ module.exports = {
         'external-handler': {
           git: require('./app/sync/external-handler/git.js')
         }
-      },
-      views: require('./app/views.js')
+      }
     },
     backlinks: {
       obs: require('./backlinks/obs.js')
