+++ lib/main-window.js
@@ -17,6 +17,7 @@ const ssbUri = require('ssb-uri')
 const pull = require('pull-stream')
 const moment = require('moment-timezone')
 const fullscreen = require('./fullscreen.js')
+const appViews = require('./app/views.js')
 
 const localTimezone = moment.tz.guess()
 moment.tz.setDefault(localTimezone)
@@ -38,7 +39,6 @@ module.exports = function (config) {
     'page.html.render': 'first',
     'app.html.search': 'first',
     'app.html.channels': 'first',
-    'app.views': 'first',
     'app.sync.externalHandler': 'first',
     'app.html.progressNotifier': 'first',
     'about.async.latestValues': 'first',
@@ -131,7 +131,7 @@ module.exports = function (config) {
     return result
   })
 
-  const views = api.app.views(api.page.html.render, defaultViews)
+  const views = appViews(api.page.html.render, defaultViews)
 
   const pendingCount = views.get('/mentions').pendingUpdates
